fixed a typo
