+++ src/main/java/org/ml_methods_group/algorithm/AKMeans.java
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package org.ml_methods_group.algoritm;
+package org.ml_methods_group.algorithm;
 
 import com.intellij.psi.PsiClass;
 import com.intellij.psi.PsiMethod;
 import com.sixrr.metrics.MetricCategory;
-import org.ml_methods_group.algoritm.entity.Entity;
+import org.ml_methods_group.algorithm.entity.Entity;
 
 import java.util.*;
