+++ src/main/java/org/ml_methods_group/algorithm/ARI.java
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package org.ml_methods_group.algoritm;
+package org.ml_methods_group.algorithm;
 
 import com.intellij.psi.PsiClass;
 import com.intellij.psi.PsiMember;
 import com.sixrr.metrics.MetricCategory;
-import org.ml_methods_group.algoritm.entity.ClassEntity;
-import org.ml_methods_group.algoritm.entity.Entity;
+import org.ml_methods_group.algorithm.entity.ClassEntity;
+import org.ml_methods_group.algorithm.entity.Entity;
 
 import java.util.*;
