+++ src/main/java/org/ml_methods_group/algorithm/RelevantProperties.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.ml_methods_group.algoritm;
+package org.ml_methods_group.algorithm;
 
 import com.intellij.psi.PsiClass;
 import com.intellij.psi.PsiField;
